requirements.txt
numpy
Pillow
SpeechRecognition
pocketsphinx
torch
torchvision
transformers
sentencepiece
accelerate
jupyter
ipykernel